Removed --network option because OpenStack API does not appear to support it
Dug through the API code and server_create code in Fog and found that the
networks get attached to the node after server_create is called, which means
there's currently no abilility to limit which networks attached by new servers.
This could be a limitation of my test environment (nova-network), but the API
docs and Fog code appear to support this conclusion.