Missing period at end of sentence.
Per our discussion that this RFC is "aspirational", I've added what I think is the desired target list
for Chef-DK platforms. The diff shows exactly the delta between where we are today and where it sounds
like we want to be (modulo my error about RHEL 5.x which isn't supported today).