Remove staff procedures
This is too heavily written around employees/staff rather than community leaders.
Perhaps a formal process is really necessary for an initial version of this, but
if we're going to lean on the Chef Software Community Team initially to handle
this we're probably best adding procedures as we work out the process for who should
handle complaints.