Update the maintenance policy to accelerate code review and merge
The current maintenance policy requires an absolute majority of maintainers to vote for a patch for it to be merged. Currently with six maintainers, this means four people must approve every patch. Adding
additional maintainers only worsens the situation and thus discourages additional involvement. Primarily this patch changes the requirement to two votes, but also adds the option of a -1 maintainer vote triggering
the requirement of a larger majority to facilitate disputes being discussed and resolved without requiring an appeal.

This patch also adds consistency in referring to components, removes the MAINTAINERS file which now exists in the Chef repository, and performs some other minor cleanup.