Set max server api version to be not configurable.
Maximum version modified to not be configurable at the server.  Having it configurable adds another setting to manage that has little benefit - if an installation of chef server is not ready for the newest features, then they would be more likely to simply not upgrade than to upgrade then take extra steps to limit new features.