Add deprecation RFC
There's still details here that need to be filled in after discussion,
but I think it's nice to break this out from the CalVer discussion. It's
similar, but different. This focuses specifically on the process of
deprecation, which we need to hammer out regardless of what version
scheme we use.

Signed-off-by: Tim Smith <tsmith@chef.io>