Exclude documentation contributions from DCO
Updating RFC 80 to eliminate the DCO requirement for commits
to documentation repositories or docs folders within projects, such
as InSpec.

chef-web-docs is already excluded from requiring DCO as stated on
the docs site (https://docs.chef.io/community_contributions.html#developer-certification-of-origin-dco)
but does not appropriately address docs embedded within a project.
This change makes the expectations clearer and consistent with the
docs.chef.io.

An additional PR to the docs site will be submitted in parallel to
this PR.

Signed-off-by: Adam Leff <adam@leff.co>