Update language about changes to interfaces
We should ensure that resources that are added to core are implemented and tested within cookbooks
prior to inclusion. A cookbook may identify the need for a particular resource to be included, but
it's fundamental implementation may have design concerns, or the resource may conflict with
standards in core. All resources should be implemented in cookbooks prior to adoption, but major
changes to the cookbook may cause issues for users of the cookbook. Instead of modifying the
cookbook, creating a new cookbook ensures that testing can be done prior to adoption without forcing
major changes to users.

Signed-off-by: Jennifer Davis <iennae@gmail.com>