tray-app
The goal of the Chef Workstation tray application is to provide a greater user experience on the desktop for Chef users.

Initially, we want to start with automatic updates as the first feature to address a problem we’ve heard for a long time from our community members and customers - the process of checking and updating versions of ChefDK and Chef Workstation isn’t as a smooth of an experience as the team wanted. Often users are left in the dark when new features or bug fixes that addressed a particular concern are made available. With automatic updates, the tray application will notify the user about new versions with a clear changelog, and facilitate the update if accepted by the user.

This will set a great foundation for the team to focus on features that will provide improved cohesion of the current Chef tooling and what’s to come.