update ADVOCATES: replace IRC with Slack, update ombudsperson
As a company we're not maintaining IRC anymore. These are the humans who
have opted to be contacts for issues in the Chef Community Slack.

Ombudsperson is updated to Mandi.

Tweaks the rake task, too, so that a useful error is raised by fetch
when a person isn't listed in the TOML.

Co-authored-by: Tim Smith <tsmith@chef.io>
Signed-off-by: Robb Kidd <robb@thekidds.org>